requirements.txt
fastapi==0.109.0
uvicorn==0.27.0
aiohttp==3.9.1
pillow==10.2.0
python-multipart==0.0.6